Relative risks for cancer from Pandeya et al

Relative risks for cancer given different levels of alcohol consumption

Usage

cancer_rr

Format

A data frame with 24 rows and 5 columns:

cancer_type

The type of cancer associated to the relative risk

dose

Dose in grams/day for which the relative risk was estimated

RR

Relative risk for cancer given the dose of alcohol

lower_CI, upper_CI

Lower and upper bounds for the 95% confidence interval

References

Pandeya, Nirmala, et al. "Cancers in Australia in 2010 attributable to the consumption of alcohol." Australian and New Zealand journal of public health 39.5 (2015): 408-413.

See also