Potential Impact Fraction related classes

Objects for handling potential impact fractions for a categorical exposure considering an observed prevalence of p and a relative risk (or relative risk parameter) of beta.

Usage

pif_class(
  pif = integer(0),
  variance = integer(0),
  conf_level = integer(0),
  type = "PIF",
  link = function() NULL,
  link_inv = function() NULL,
  link_deriv = function() NULL
)

pif_atomic_class(
  p,
  p_cft,
  beta,
  var_p,
  var_beta,
  rr_link,
  rr_link_deriv,
  link,
  link_deriv,
  link_inv,
  conf_level,
  type,
  upper_bound_p,
  upper_bound_beta
)

pif_total_class(
  pif_list,
  weights,
  sigma_weights,
  conf_level = 0.95,
  link,
  link_inv,
  link_deriv
)

pif_ensemble_class(pif_list, conf_level = 0.95)

Arguments

pif

Potential Impact Fraction estimate

variance

variance estimate for the potential impact fraction (i.e. for pif)

conf_level

Confidence level for the confidence interval (default 0.95).

type

Character either Potential Impact Fraction (PIF) or Population Attributable Fraction (PAF)

link

Link function such that the pif confidence intervals stays within the expected bounds.

link_inv

The inverse of link. For example if link is logit this should be inv_logit.

link_deriv

The derivative of link. For example if link is logit this should be deriv_logit (i.e. function(pif) 1 / (pif * (1 - pif))).

p

Prevalence (proportion) of the exposed individuals for each of the N exposure levels.

p_cft

Counterfactual prevalence (proportion) of the exposed individuals for each of the N exposure levels.

beta

Relative risk parameter for which standard deviation is available (usually its either the relative risk directly or the log of the relative risk as most RRs, ORs and HRs come from exponential models).

var_p

Estimate of the colink_variance matrix of p where the entry var_p[i,j] represents the colink_variance between p[i] and p[j].

var_beta

Estimate of the colink_variance matrix of beta where the entry var_beta[i,j] represents the colink_variance between beta[i] and beta[j].

rr_link

Link function such that the relative risk is given by rr_link(beta).

rr_link_deriv

Derivative of the link function for the relative risk. The constructor tries to build it automatically from rr_link using Deriv::Deriv().

upper_bound_p

Whether the values for the p component of the link_variance should be approximated by an upper bound.

upper_bound_beta

Whether the values for the beta component of the link_variance should be approximated by an upper bound.

pif_list

A list of potential impact fractions pif_class so that the total can be computed from it.

weights

Weights for calculating the total PIF (respectively PAF) in pif_total.

sigma_weights

Colink_variance matrix for the weights when calculating the total PIF (respectively PAF) in pif_total.

Properties of a pif_class

ci

numeric(2) — Lower and upper confidence limits at level conf_level.

link_vals

numeric — Entrywise evaluation of the link function at pif: link(pif).

link_deriv_vals

character — Entrywise evaluation of the derivative of the link function (link_deriv) at pif: link(pif).

link_variance

numeric - Estimate for the linked potential impact fraction's variance: var(link(pif)).

Properties of a pif_atomic_class

The pif_atomic_class inherits the properties of a pif_class as well as:

mu_obs

numeric — Average relative risk in the observed population.

mu_cft

numeric — Average relative risk in the counterfactual population.

pif

numeric — Estimate of the potential impact fraction.

rr_link_deriv_vals

character — Entrywise evaluation of the derivative of the link function (link_deriv) at pif: link(pif).

Computation of confidence intervals

Wald-type confidence intervals are calculated for link(pif) as follows: $$ \text{CI}_{\text{Link}} = \text{link}\big(\text{PIF}\big) \pm Z_{\alpha/2}\cdot\sqrt{\textrm{link\_variance}} $$ and then transformed back using the inverse of the link function inv_link: $$ \text{CI}_{\text{PIF}} = \text{link}^{-1}\Big(\text{CI}_{\text{Link}}\Big) $$

Class structure

All potential impact fractions inherit from the pif_class which provides some of the generics.

The pif_atomic_class is a type of pif_class that contains enough information for it to be computed through the classic formula by Walter: $$ \textrm{PIF} = \dfrac{ \sum\limits_{i=1}^N p_i \text{RR}_i - \sum\limits_{i=1}^N p_i^{\text{cft}} \text{RR}_i }{ \sum\limits_{i=1}^N p_i \text{RR}_i } $$ where the relative risk is a function of a parameter \(\beta_i\) $$ \text{RR}_i = g(\beta_i) $$ and the link_variance is calculated for a function of PIF: \(f(\textrm{PIF})\) The pif_atomic_class only contains one potential impact fraction and the parameters to estimate it.

The pif_additive_class is a type of pif_class that can be computed as a sum of weighted transformations of potential impact fractions where the weights can be random. Elements of a pif_additive_class have a link_variance estimated for the following expression: $$ f\big(\textrm{PIF}_{+}\big) = \sum\limits_{i = 1}^{N} q_i \cdot f_i\Big(\textrm{PIF}_i\Big) $$

Examples of calculations that can be added to a pif_additive_class are:Ç

The total potential impact fraction (combining different subpopulations) $$ \textrm{PIF}_{Total} = \sum\limits_{i = 1}^{N} q_i \cdot \textrm{PIF}_i $$ with \(q_i\) representing the proportions of individuals in each category.

The ensemble potential impact fraction (representing different relative risks) for the same outcome $$ \textrm{PIF}_{Ensemble} = 1 - \prod\limits_{i = 1}^{N} \Big(1 - \textrm{PIF}_i\Big) $$ as it can be transformed into $$ \ln\Big(1 - \textrm{PIF}_{Ensemble}\Big) = \sum\limits_{i = 1}^{N} \ln\big(1 - \textrm{PIF}_i\big) $$

Examples

#Create a new pif parent class element
pif_class(pif = 0.3, var = 0.01, conf_level = 0.95, type = "PIF",
  link = logit, link_inv = inv_logit, link_deriv = deriv_logit)
#> 
#> ── Potential Impact Fraction ──
#> 
#> PIF = 30.000% [95% CI: 14.423% to 52.149%]
#> sd(pif %) = 10.000
#> sd(link(pif)) = 0.476

#Create a new potential impact fraction from the Walter's formula
pif_atomic_class(
  p = 0.499, p_cft = 0, beta = 3.6, var_p = 0.1, var_beta = 3,
  link = logit, link_inv = inv_logit, link_deriv = deriv_logit,
  rr_link = identity, rr_link_deriv = function(x) 1,
  conf_level = 0.95, type = "PAF",
  upper_bound_p = FALSE,
  upper_bound_beta = FALSE
)
#> 
#> ── Population Attributable Fraction ──
#> 
#> PAF = 56.473% [95% CI: 17.628% to 88.720%]
#> sd(paf %) = 22.601
#> sd(link(paf)) = 0.919

#Create a list of pif
pif1 <- pif_atomic_class(
  p = 0.499, p_cft = 0, beta = 3.6, var_p = 0.01, var_beta = 0.03,
  link = logit, link_inv = inv_logit, link_deriv = deriv_logit,
  rr_link = identity, rr_link_deriv = function(x) 1,
  conf_level = 0.95, type = "PAF",
  upper_bound_p = FALSE,
  upper_bound_beta = FALSE
)
pif2 <- pif_atomic_class(
  p = 0.79, p_cft = 0, beta = 3.6, var_p = 0.01, var_beta = 0.03,
  link = logit, link_inv = inv_logit, link_deriv = deriv_logit,
  rr_link = identity, rr_link_deriv = function(x) 1,
  conf_level = 0.95, type = "PAF",
  upper_bound_p = FALSE,
  upper_bound_beta = FALSE
)
pif3 <- pif_atomic_class(
  p = 0.8, p_cft = 0, beta = 3.6, var_p = 0.01, var_beta = 0.03,
  link = logit, link_inv = inv_logit, link_deriv = deriv_logit,
  rr_link = identity, rr_link_deriv = function(x) 1,
  conf_level = 0.95, type = "PAF",
  upper_bound_p = FALSE,
  upper_bound_beta = FALSE
)

tp1 <- pif_total_class(pif_list = list(pif1, pif2),
  weights = c(0.5, 0.2), sigma_weights = diag(0.001, ncol = 2, nrow = 2),
  link = identity, link_inv = identity, link_deriv = identity)

pif_total_class(pif_list = list(tp1, pif3),
  weights = c(0.7, 0.3), sigma_weights = diag(0.001, ncol = 2, nrow = 2),
  link = identity, link_inv = identity, link_deriv = identity)
#> 
#> ── Population Attributable Fraction ──
#> 
#> PAF = 49.441% [95% CI: 45.514% to 53.368%]
#> sd(paf %) = 4.052
#> sd(link(paf)) = 0.020