Covariance between two atomic pifs

Calculates the approximate covariance between two atomic pif class expressions using the delta method.

Usage

cov_atomic_pif(pif1, pif2, var_p, var_beta)

Arguments

pif1

A pif_atomic_class object.

pif2

A second pif_atomic_class object.

var_p

covariance matrix for the prevalences in both pif1 and pif2. Default NULL (no correlation).

var_beta

covariance matrix for the parameter beta in both pif1 and pif2. Default NULL (no correlation).

Computation

The cov_atomic_pif computes the approximate covariance between two potential impact fractions \(\text{PIF}_1\) and \(\text{PIF}_2\) as follows: $$ \text{Cov}\Big(\text{PIF}_1, \text{PIF}_2 \Big) \approx \Bigg( \dfrac{\partial \text{PIF}_1}{\partial p_1}^{\top} \Sigma_p \dfrac{\partial \text{PIF}_2}{\partial p_2} + \dfrac{\partial \text{PIF}_1}{\partial \beta_1}^{\top} \Sigma_{\beta} \dfrac{\partial \text{PIF}_2}{\partial \beta_2} \Bigg) $$ where \(\Sigma_p\) is var_p and \(\Sigma_{\beta}\) is var_beta. The parameters \(p_1\) and \(\beta_1\) refer to the prevalence (p) and relative risk exposure parameter beta of the first potential impact fraction pif1 and \(p_2\) and \(\beta_2\) to the respective parameters of pif2.

See also