Covariance between pif_global_ensemble_class and a pif_atomic

Calculates the covariance of a potential impact fraction of a pif_global_ensemble_class with a pif_atomic

Usage

cov_ensemble_atomic(
  pif_ensemble,
  pif_atomic,
  var_p,
  var_beta,
  var_pifs,
  var_pif_weights,
  recursive = TRUE
)

Arguments

pif_ensemble

A pif_global_ensemble_class

pif_atomic

A pif_atomic_class

var_pifs

Covariance vector between the potential impact fractions in pif_ensemble and pif_atomic.

var_pif_weights

Covariance vector between the weights in pif_ensemble and the pif_atomic.

Formula

Given a pif_global_ensemble: $$ \widehat{\textrm{PIF}}_{1} = g^{-1}\Bigg(\sum\limits_{i=1}^{M_1} g(\hat{q}_i \cdot \widehat{\textrm{PIF}}_{1,i})\Bigg) $$ and a second pif_global_ensemble: $$ \widehat{\textrm{PIF}}_{2} = h^{-1}\Bigg(\sum\limits_{j=1}^{M_2} g(\hat{w}_j \cdot \widehat{\textrm{PIF}}_{2,j})\Bigg) $$ This function computes the covariance between the first impact fraction and the coefficients of the second fraction by computing: $$ \operatorname{Cov}(\widehat{\textrm{PIF}}_A, \widehat{\textrm{PIF}}_{B,j}) \approx \frac{1}{g'\big(\widehat{\textrm{PIF}}_A\big)}\sum\limits_{i=1}^{M_1}g'(\hat{q}_i \cdot \widehat{\textrm{PIF}}_{A,i}) \Bigg[ \mathbb{E}[\hat{q}_i] \operatorname{Cov}\Big( \textrm{PIF}_{A,i}, \widehat{\textrm{PIF}}_{B,j}\Big) + \mathbb{E}[\textrm{PIF}_{A,i}] \operatorname{Cov}\Big( \hat{q}_i, \widehat{\textrm{PIF}}_{B,j}\Big) \Bigg] $$ where \(\widehat{\textrm{PIF}}_{1,:} = (\widehat{\textrm{PIF}}_{1,1}, \widehat{\textrm{PIF}}_{1,2}, \dots, \widehat{\textrm{PIF}}_{1,M_1})^{\top}\)

See also