Covariance between the j-th weight of a pif_global_ensemble_class and another pif_global_ensemble_class

Calculates the covariance of the jth-weight of a potential impact fraction of pif_global_ensemble_class with a second pif_global_ensemble_class or pif_atomic_class.

Usage

cov_ensemble_weights(
  pif1,
  pif2,
  var_weights,
  var_pif_weights,
  recursive = !is.null(var_pif_weights)
)

Arguments

pif1

A pif_global_ensemble_class from which the weight is taken.

pif2

A pif_global_ensemble_class or pif_atomic_class to compute the covariance

var_weights

Covariance vector between the weights of pif1 and the j-th weight of pif2.

var_pif_weights

Covariance structure between the potential impact fractions in pif1 and the weights in pif2.

Note

The model currently works under the assumption that all pif_atomic_class are independent from weights. Hence will return 0 if pif2 is an atomic pif.

Formula

Given a pif_global_ensemble: $$ \widehat{\textrm{PIF}}_{1} = g^{-1}\Bigg(\sum\limits_{i=1}^{M_1} g(\hat{q}_i \cdot \widehat{\textrm{PIF}}_{1,i})\Bigg) $$ and a second pif_global_ensemble: $$ \widehat{\textrm{PIF}}_{2} = h^{-1}\Bigg(\sum\limits_{j=1}^{M_2} g(\hat{w}_j \cdot \widehat{\textrm{PIF}}_{2,j})\Bigg) $$ This function computes the covariance between the first impact fraction and the jth weight \(w_j\) of the second fraction by computing: $$ \operatorname{Cov}(\widehat{\textrm{PIF}}_A, \hat{w}_j) \approx \frac{1}{g'\big(\widehat{\textrm{PIF}}_A\big)}\sum\limits_{i=1}^{M_1}g'(\hat{q}_i \cdot \widehat{\textrm{PIF}}_{A,i}) \Bigg[ \hat{q} \operatorname{Cov}\Big( \widehat{\textrm{PIF}}_{A,i}, \hat{w}_j\Big) + \widehat{\textrm{PIF}}_{A,i} \operatorname{Cov}\Big( \hat{q}_j, \hat{w}_j\Big) \Bigg] $$ where \(\widehat{\textrm{PIF}}_{1,:} = (\widehat{\textrm{PIF}}_{1,1}, \widehat{\textrm{PIF}}_{1,2}, \dots, \widehat{\textrm{PIF}}_{1,M_1})^{\top}\)

See also