Covariance function for a pif_global_ensemble_class

Recursively obtains the covariance matrix of a pif_global_ensemble_class

Usage

cov_total_pif(
  pif1,
  pif2,
  var_p = NULL,
  var_beta = NULL,
  var_weights = NULL,
  var_pif_weights = NULL,
  var_pifs = NULL,
  warning = TRUE
)

Arguments

pif1

Either a pif_atomic_class or a pif_global_ensemble_class

pif2

Either a pif_atomic_class or a pif_global_ensemble_class

var_p

covariance matrix for the prevalences in both pif1 and pif2. Default NULL (no correlation).

var_beta

covariance matrix for the parameter beta in both pif1 and pif2. Default NULL (no correlation).

var_weights

Covariance between the weights of all the fractions contained in pif1 and all the fractions contained in pif2.

var_pif_weights

Covariance vector between the weights in pif_ensemble and the pif_atomic. This refers to the term \(\operatorname{Cov}\Big( \hat{q}_i,\widehat{\textrm{PIF}}_{B,j}\Big)\) in the equation below. If set to NULL its automatically calculated.

var_pifs

Covariance vector between the potential impact fractions in pif_ensemble and pif_atomic. This refers to the term \(\operatorname{Cov}\Big( \textrm{PIF}_{A,i}, \widehat{\textrm{PIF}}_{B,j}\Big)\) in the equation below. If set to NULL its automatically calculated.

warning

Boolean indicating whether to throw a warning if the labels on the fractions involved are not unique.

Computation

This computes: $$ \operatorname{Cov}(\widehat{\textrm{PIF}}_{A}, \widehat{\textrm{PIF}}_{B}) \approx \frac{1}{g'\big(\widehat{\textrm{PIF}}_A\big)}\sum\limits_{i=1}^{M_1}g'(\hat{q}_i \cdot \widehat{\textrm{PIF}}_{A,i}) \Bigg[ \mathbb{E}[\hat{q}_i] \operatorname{Cov}\Big( \textrm{PIF}_{A,i}, \widehat{\textrm{PIF}}_{B}\Big) + \mathbb{E}[\textrm{PIF}_{A,i}] \operatorname{Cov}\Big( \hat{q}_i, \widehat{\textrm{PIF}}_{B}\Big) \Bigg] $$ where \(\operatorname{Cov}\Big( \textrm{PIF}_{A,i}, \widehat{\textrm{PIF}}_{B}\Big)\) is computed as in cov_ensemble_atomic() in the case \(\widehat{\textrm{PIF}}_{A,i}\) is atomic or with the same formula in the case it is an ensemble. The expression \(\operatorname{Cov}\Big( \hat{q}_i, \widehat{\textrm{PIF}}_{B}\Big) \) is computed as in cov_ensemble_weights() for each weight \(q_j\).

See also