Covariance component with respect to p

Calculates the covariance p or beta component between two potential impact fractions or the complete covariance.

Usage

from_parameters_covariance_p_component(
  p1,
  p2,
  p1_cft,
  p2_cft,
  rr1,
  rr2,
  mu_obs1,
  mu_obs2,
  mu_cft1,
  mu_cft2,
  var_p,
  upper_bound
)

from_parameters_covariance_beta_component(
  p1,
  p2,
  p1_cft,
  p2_cft,
  rr1,
  rr2,
  rr_link_deriv_vals1,
  rr_link_deriv_vals2,
  mu_obs1,
  mu_obs2,
  mu_cft1,
  mu_cft2,
  var_beta,
  upper_bound
)

from_parameters_pif_covariance(
  p1,
  p2,
  p1_cft,
  p2_cft,
  rr1,
  rr2,
  rr_link_deriv_vals1,
  rr_link_deriv_vals2,
  mu_obs1,
  mu_obs2,
  mu_cft1,
  mu_cft2,
  var_p,
  var_beta,
  upper_bound_p = FALSE,
  upper_bound_beta = FALSE
)

from_parameters_pif_variance(
  p,
  p_cft,
  rr,
  rr_link_deriv_vals,
  mu_obs,
  mu_cft,
  var_p,
  var_beta,
  upper_bound_p = FALSE,
  upper_bound_beta = FALSE
)

Arguments

p1

Observed proportion exposed in the first potential impact fraction

p2

Observed proportion exposed in the first potential impact fraction

p1_cft

Counterfactual proportion exposed in the first potential impact fraction

p2_cft

Counterfactual proportion exposed in the first potential impact fraction

rr1

Relative risk in the first potential impact fraction

rr2

Relative risk in the second potential impact fraction

mu_obs1

Average relative risk in the population of the first potential impact fraction.

mu_obs2

Average relative risk in the population of the second potential impact fraction.

mu_cft1

Average relative risk in the counterfactual population of the first potential impact fraction.

mu_cft2

Average relative risk in the counterfactual of the second potential impact fraction.

var_p

Covariance matrix with the entry var_p[i,j] corresponding to the covariance between p1[i] and p2[j].

upper_bound

Whether the variance should be calculated or an upper bound for the variance (assuming perfect correlation).

rr_link_deriv_vals1

Values for the derivative of the relative risk function evaluated at theta (the relative risk parameter).

rr_link_deriv_vals2

Values for the derivative of the relative risk function evaluated at theta (the relative risk parameter).

var_beta

Estimate of the colink_variance matrix of beta where the entry var_beta[i,j] represents the colink_variance between beta[i] and beta[j].

upper_bound_p

Whether the values for the p component of the link_variance should be approximated by an upper bound.

upper_bound_beta

Whether the values for the beta component of the link_variance should be approximated by an upper bound.

p

Prevalence (proportion) of the exposed individuals for each of the N exposure levels.

p_cft

Counterfactual prevalence (proportion) of the exposed individuals for each of the N exposure levels.

rr

The relative risk for each of the exposure levels.

rr_link_deriv_vals

The derivative of the relative risk function g with respect to the parameter beta evaluated at beta.

mu_obs

The average value of the relative risk in the observed population.

mu_cft

The average value of the counterfactual relative risk in the population.

Value

The covariance component for either p or beta

Note

This estimation does not require a pif_class object but instead is built from parameters.

Formulas

The following represents the covariance components. For p:

$$ \text{CC}_{p} = \dfrac{ \partial \textrm{PIF}_i}{\partial p_i}^{\top} \textrm{covariance}\big( \hat{p}_i, \hat{p}_j\big) \dfrac{\partial \textrm{PIF}_j}{\partial p_j} $$

and for beta:

$$ \text{CC}_{\theta} = \dfrac{ \partial \textrm{PIF}_i}{\partial \theta_i}^{\top} \textrm{covariance}\big( \hat{\theta}_i, \hat{\theta}_j\big) \dfrac{\partial \textrm{PIF}_j}{\partial \theta_j} $$

See also