Getters

Collection of functions for the potential impact fraction class so that they work as getters of the properties. Each function is constructed as get_property

Usage

get_rr(self)

get_mu_obs(self)

get_mu_cft(self)

get_pif(self)

get_link_vals(self)

get_link_deriv_vals(self)

get_rr_link_deriv_vals(self)

get_link_variance(self)

get_variance_atomic(self)

get_ci(self)

get_total_coefs(self)

get_ensemble_coefs(self)

get_ensemble_pif(self)

get_total_pif(self)

get_total_type(self)

get_ensemble_type(self)

get_covariance_total(self)

get_ensemble_covariance(self)

get_variance_total(self)

get_ensemble_variance(self)

Arguments

self

A pif_class or pif_atomic_class object created with S7.