Get the colink_variances between two pif classes (atomic pifs)

Get the colink_variances between two pif classes (atomic pifs)

Usage

pif_class_atomic_variance(
  pif1,
  pif2,
  var_p = NULL,
  var_beta = NULL,
  independent_p = "guess",
  independent_beta = "guess",
  quiet = FALSE
)

Arguments

pif1

A pif_class object

pif2

A pif_class object

var_p

Colink_variance matrix for the prevalences in both pif1 and pif2.

var_beta

Colink_variance matrix for the parameter beta in both pif1 and pif2.

independent_p

If pif1 and pif2 share the same prevalence data. Either TRUE, FALSE or guess (default).

independent_beta

If pif1 and pif2 share the same beta parameter. Either TRUE, FALSE or guess (default).

quiet

Whether to throw warnings or other messages

Colink_variance matrices

By default if no var_p is specified this assumes the colink_variances between the parameters p of pif1 and pif2 are uncorrelated. However, if pif1 and pif2 share the same prevalence estimates (i.e. share the same ps, then the user should set independent_p = TRUE to account for that correlation).

The same thing happens with var_beta. If no var_beta is specified then the assumption is that the beta parameters from pif1 and from pif2 are uncorrelated unless independent_beta is set to TRUE.

If the user provides a colink_variance matrix for var_p then independent_p is disregarded. Similarly, if the user provides var_beta then independent_beta is ignored.