Combine Potential Impact Fractions and Population Attributable Fractions

Combine potential impact fractions or the population attributable fractions to either generate the total fraction from the fractions of subpopulations (pif_total/paf_total) or the ensemble fraction of a population from different (independent) exposures.

Usage

paf_total(
  paf1,
  ...,
  weights,
  sigma_weights = 0,
  conf_level = 0.95,
  link = "log-complement",
  link_inv = NULL,
  link_deriv = NULL,
  quiet = FALSE
)

pif_total(
  pif1,
  ...,
  weights,
  sigma_weights = 0,
  conf_level = 0.95,
  link = "log-complement",
  link_inv = NULL,
  link_deriv = NULL,
  quiet = FALSE
)

pif_ensemble(pif1, ..., conf_level = 0.95, quiet = FALSE)

Arguments

paf1

A population attributable fraction (class pif_class)

...

The remaining potential impact fractions or (respectively) population attributable fractions.

weights

A vector containing the proportion of the population for each of the categories (for each of the pifs given).

sigma_weights

Colink_variance structure for the weights. Can be 0 (default) if the weights are not random, a vector if only the link_variances of the weights are available or a colink_variance matrix.

conf_level

Confidence level for the confidence interval (default 0.95).

link

Link function such that the pif confidence intervals stays within the expected bounds.

link_inv

The inverse of link. For example if link is logit this should be inv_logit.

link_deriv

Derivative of the link function. The function tries to build it automatically from link using Deriv::Deriv().

quiet

Whether to show messages.

pif1

A potential impact fraction (class pif_class)

Total potential impact fraction

Assuming the overall population can be subdivided into \(N\) distinct subpopulations each of them with a different potential impact fraction (or population attributable fraction) we can estimate the total population attributable fraction or potential impact fraction of the whole population as: $$ \text{PIF}_{\text{Total}} = \sum\limits_{i = 1}^{N} \pi_i \cdot \text{PIF}_i $$ where each \(\text{PIF}_i\) corresponds to the potential impact fraction of the i-th subpopulation and \(\pi_i\) correspond to the proportion of the total population occupied by \(\text{PIF}_i\). The weights are such that \(\sum_{i=1}^{N} \pi_i = 1\).

Ensemble potential impact fraction

If a population is exposed to \(K\) different independent risk factors then the ensemble impact fraction of the combination of those factors can be written as: $$ \text{PIF}_{\text{Ensemble}} = 1 - \prod\limits_{\ell = 1}^{K} \Big(1 - \cdot \text{PIF}_{\ell}\Big) $$ where each \(\text{PIF}_{\ell}\) corresponds to the potential impact fraction of the \(\ell\)-th risk factor for the same population.

Examples

#Potential impact fraction for women
pif_women <- pif(0.32, 0.1, 1.2, quiet = TRUE, var_p = 0.1)

#Potential impact fraction for men
pif_men <- pif(0.27, 0.1, 1.3, quiet = TRUE, var_p = 0.1)

#Population potential impact fraction with 49% men and 51% women
pif_total(pif_men, pif_women, weights = c(0.49, 0.51), link = "logit")
#> 
#> ── Potential Impact Fraction ──
#> 
#> PIF = 4.421% [95% CI: 0.179% to 54.364%]
#> sd(pif %) = 7.003
#> sd(link(pif)) = 1.657

#Population attributable  fraction for women
paf_women <- paf(0.32, 1.3, quiet = TRUE, var_p = 0.1)

#Population attributable  fraction for men
paf_men <- paf(0.27, 1.3, quiet = TRUE, var_p = 0.1)
paf_total(paf_men, paf_women, weights = c(0.49, 0.51), link = "logit")
#> 
#> ── Population Attributable Fraction ──
#> 
#> PAF = 8.139% [95% CI: 1.075% to 41.949%]
#> sd(paf %) = 8.006
#> sd(link(paf)) = 1.071

# Calculate the ensemble from lead and radiation exposure
paf_lead <- paf(0.2, 2.2, quiet = TRUE, var_p = 0.001)
paf_rad  <- paf(0.1, 1.2, quiet = TRUE, var_p = 0.0001)
pif_ensemble(paf_lead, paf_rad)
#> 
#> ── Population Attributable Fraction ──
#> 
#> PAF = 0.380% [95% CI: -7.002% to 7.252%]
#> sd(paf %) = 3.633
#> sd(link(paf)) = 0.036

# Totals and ensembles can be combined
pif_lead_women <- paf(0.27, 2.2, quiet = TRUE, var_p = 0.001)
pif_rad_women  <- paf(0.12, 1.2, quiet = TRUE, var_p = 0.001)
pif_women      <- pif_ensemble(pif_lead_women, pif_rad_women)
pif_lead_men   <- paf(0.30, 2.2, quiet = TRUE, var_p = 0.001)
pif_rad_men    <- paf(0.10, 1.2, quiet = TRUE, var_p = 0.001)
pif_men        <- pif_ensemble(pif_lead_men, pif_rad_men)

pif_total(pif_men, pif_women, weights = c(0.49, 0.51))
#> 
#> ── Population Attributable Fraction ──
#> 
#> PAF = 0.547% [95% CI: -4.919% to 5.728%]
#> sd(paf %) = 2.715
#> sd(link(paf)) = 0.027