Get the covariance between two atomic pifs

Calculates the approximate covariance between two atomic pif class expressions using the delta method.

Usage

cov_atomic_pif(
  pif1,
  pif2,
  var_p = NULL,
  var_beta = NULL,
  uncorrelated_p = "guess",
  uncorrelated_beta = "guess",
  quiet = FALSE
)

Arguments

pif1

A pif_atomic_class object.

pif2

A second pif_atomic_class object.

var_p

covariance matrix for the prevalences in both pif1 and pif2. Default NULL (no correlation).

var_beta

covariance matrix for the parameter beta in both pif1 and pif2. Default NULL (no correlation).

uncorrelated_p

If pif1 and pif2 were constructed with independent (uncorrelated) data for the prevalence p. Either TRUE, FALSE or guess (default).

uncorrelated_beta

If pif1 and pif2 were constructed with independent (uncorrelated) data for the relative risk parameter beta. Either TRUE, FALSE or guess (default).

quiet

Whether to throw warnings or other messages.

Link covariance matrices

By default if var_p = NULL is specified this assumes the parameters p of pif1 and pif2 are uncorrelated. However, if pif1 and pif2 share the same prevalence estimates (i.e. share the same ps, then the user should set uncorrelated_p = FALSE to account for that correlation).

The same thing happens with var_beta. If no var_beta is specified (set to NULL) then the assumption is that the beta parameters from pif1 and from pif2 are uncorrelated (unless uncorrelated_beta is set to FALSE).

If the user provides a covariance matrix for var_p then uncorrelated_p is disregarded. Similarly, if the user provides var_beta then uncorrelated_beta is ignored.

Computation

The pif_class_atomic_variance computes the approximate covariance between two potential impact fractions \(\text{PIF}_1\) and \(\text{PIF}_2\) as follows: $$ \text{Cov}\Big(\text{PIF}_1, \text{PIF}_2 \Big) \approx \Bigg( \dfrac{\partial \text{PIF}_1}{\partial p_1}^{\top} \Sigma_p \dfrac{\partial \text{PIF}_2}{\partial p_2} + \dfrac{\partial \text{PIF}_1}{\partial \beta_1}^{\top} \Sigma_{\beta} \dfrac{\partial \text{PIF}_2}{\partial \beta_2} \Bigg) $$ where \(\Sigma_p\) is var_p and \(\Sigma_{\beta}\) is var_beta. The parameters \(p_1\) and \(\beta_1\) refer to the prevalence (p) and relative risk exposure parameter beta of the first potential impact fraction pif1 and \(p_2\) and \(\beta_2\) to the respective parameters of pif2.

See also